* 'master' of git@gitlab.com:wireshark/wireshark.git: (606 commits)
  rc-v3: Deduplicate fields
  pkix1explicit: Deduplicate some incompatible field names
  file-rtpdump: Fix duplicated field name.
  IRC: Use proto_irc isf hf_irc_ctcp(Duplicated fields)
  rtp-ed137: Fix duplicated fields warning.
  Revert "poc-settings-dtd:poc-settings' exists multiple times with incompatible types"
  wustil: Add David Gay's dtoa for shortest sufficient double representation
  Optimize hash table usage in dissectors
  GNSS: dissect Galileo SAR Short-RLM
  TCP: Add an analysis field for Karn-Partridge ambiguous ACKs
  docs: Update the Stratoshark quick start
  IEEE1722: Don't reuse can.reserved in ACF CAN
  packet-communityid: Fix duplicated field name.
  poc-settings-dtd:poc-settings' exists multiple times with incompatible types
  MQ: conflicting entry in value_string (MQCA_APPL_ID) and (MQCA_FIRST)
  ORAN FH CUS: Show meaning of symbolMask for ASM command
  packet-delep: Fix duplicated filter ids.
  packet-asam-cmp: Fix duplicated field name.
  show_exception: Remove duplicated field warning.
  HICP: Allow filtering the error string
  ...

# Conflicts:
#	.github/workflows/macos.yml
#	.github/workflows/ubuntu.yml

+++ capture/capture_session.h
@@ -17,7 +17,7 @@
 #include <stdint.h>
 #endif
 
-#include "capture_opts.h"
+#include "ui/capture_opts.h"
 
 #include <epan/fifo_string_cache.h>
 #include <wsutil/processes.h>
@@ -99,7 +99,7 @@ struct _capture_session {
     gid_t     group;                      /**< group of the cfile */
 #endif
     bool  session_will_restart;       /**< Set when session will restart */
-    uint32_t   count;                      /**< Total number of frames captured */
+    uint64_t   count;                      /**< Total number of frames captured */
     uint32_t   count_pending;              /**< Number of frames captured but not yet read */
     capture_options *capture_opts;        /**< options for this capture */
     capture_file *cf;                     /**< handle to cfile */
